pyproject.toml
[project]
name = "voice-assistant-livekit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask[async]>=3.1.1",
    "flask-cors>=6.0.1",
    "livekit-agents[groq]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
    "livekit-api>=1.0.3",
    "livekit-plugins-silero>=1.1.3",
]
